src/core/fileParser.py
from PySide6.QtCore import Signal, QThread

class FileParserWorker(QThread):
    """Worker thread for log parsing"""
    progress_updated = Signal(int, str)
    parsing_complete = Signal(list, dict)
    
    def __init__(self, log_files, patterns):
        super().__init__()
        self.log_files = log_files
        self.patterns = patterns
    
    def run(self):
        # Integrate your existing parsing logic here
        # Emit progress signals during processing
        pass